requirements.txt
newspaper3k
beautifulsoup4
lxml
langdetect
streamlit
python-dotenv
requests
pyttsx3